Move RawTask to internal.
