+++ src/internal/system/schedule/mod.rs
@@ -1,1 +1,2 @@
+pub(crate) mod raw_task;
 pub(crate) mod task;
