+++ src/public/system/schedule/builder.rs
@@ -1,120 +1,22 @@
 use crate::{
     internal::{
-        query::claim::Claim,
         system,
-        system::schedule::task::Task,
+        system::schedule::{task::Task, raw_task::{RawTasks, RawTask}, raw_task},
     },
     system::{
-        schedule::{
-            stage,
-            stage::{Stage, Stages},
-            Schedule,
-        },
+        schedule::Schedule,
         ParSystem, System,
     },
 };
-use core::any::TypeId;
 use hashbrown::HashSet;
-
-pub enum RawTask<S, P> {
-    Task(Task<S, P>),
-    Flush,
-}
-
-pub struct Null;
-
-pub trait RawTasks<'a> {
-    type Stages: Stages<'a>;
-
-    fn into_stages(
-        self,
-        mutable_claims: &mut HashSet<TypeId>,
-        immutable_claims: &mut HashSet<TypeId>,
-        mutable_buffer: &mut HashSet<TypeId>,
-        immutable_buffer: &mut HashSet<TypeId>,
-    ) -> Self::Stages;
-}
-
-impl<'a> RawTasks<'a> for Null {
-    type Stages = stage::Null;
-
-    fn into_stages(
-        self,
-        _mutable_claims: &mut HashSet<TypeId>,
-        _immutable_claims: &mut HashSet<TypeId>,
-        _mutable_buffer: &mut HashSet<TypeId>,
-        _immutable_buffer: &mut HashSet<TypeId>,
-    ) -> Self::Stages {
-        stage::Null
-    }
-}
-
-impl<'a, S, P, T> RawTasks<'a> for (RawTask<S, P>, T)
-where
-    S: System<'a> + Send,
-    P: ParSystem<'a> + Send,
-    T: RawTasks<'a>,
-{
-    type Stages = (Stage<S, P>, T::Stages);
-
-    fn into_stages(
-        self,
-        mutable_claims: &mut HashSet<TypeId>,
-        immutable_claims: &mut HashSet<TypeId>,
-        mutable_buffer: &mut HashSet<TypeId>,
-        immutable_buffer: &mut HashSet<TypeId>,
-    ) -> Self::Stages {
-        let prev_stages = self.1.into_stages(
-            mutable_claims,
-            immutable_claims,
-            mutable_buffer,
-            immutable_buffer,
-        );
-
-        match self.0 {
-            RawTask::Task(task) => {
-                mutable_buffer.clear();
-                immutable_buffer.clear();
-
-                // Identify this stage's claims on components.
-                S::Views::claim(mutable_buffer, immutable_buffer);
-                P::Views::claim(mutable_buffer, immutable_buffer);
-
-                // Helper function to check whether the intersection betwen two sets is nonempty.
-                fn intersects(a: &HashSet<TypeId>, b: &HashSet<TypeId>) -> bool {
-                    a.intersection(b).next().is_some()
-                }
-
-                // If the claims are incompatible, a new stage must begin.
-                //
-                // Claims are incompatible if an immutable claim is made on a component already
-                // mutable claimed, or if a mutable claim is made on a component already claimed at
-                // all.
-                if intersects(immutable_buffer, mutable_claims)
-                    || intersects(mutable_buffer, mutable_claims)
-                    || intersects(mutable_buffer, immutable_claims)
-                {
-                    (Stage::Start(task), prev_stages)
-                } else {
-                    (Stage::Continue(task), prev_stages)
-                }
-            }
-            RawTask::Flush => {
-                mutable_claims.clear();
-                immutable_claims.clear();
-                (Stage::Flush, prev_stages)
-            }
-        }
-    }
-}
 
 pub struct Builder<T> {
     raw_tasks: T,
 }
 
-impl Builder<Null> {
+impl Builder<raw_task::Null> {
     pub(super) fn new() -> Self {
-        Self { raw_tasks: Null }
+        Self { raw_tasks: raw_task::Null }
     }
 }
